If you change the repo URL for an existing pipeline...
then it looks like you need to at least re-run the pipeline manually
once in order for jenkins to realize that your repo has changed (from
g0t4 in this case to aggroskater). I also clobbered the entire
workspace/atmosphere-pipeline/ directory structure on the server. Not
sure if that mattered or not, because enabling git polling after doing
that still had the poller trying to talk to g0t4 and not aggroskater.
I'm guessing it's just looking at last successful build? Or last build
in general?

Regardless, this commit should trigger another build, now that the scm
poller is looking in the aggroskater repo.